// src/lib/types/CMS.ts
export interface User {
	id: string;
	name: string;
	sections: ('hefw' | 'timelinerys' | 'dokomi-fan-booth' | 'aqua-sendoff')[];
	roles: (
		| 'superadmin'
		| 'project-owner'
		| 'content-moderator'
		| 'developer'
		| 'translator'
		| 'misc'
	)[];
	updatedAt: string;
	createdAt: string;
	enableAPIKey?: boolean | null;
	apiKey?: string | null;
	apiKeyIndex?: string | null;
	email: string;
	resetPasswordToken?: string | null;
	resetPasswordExpiration?: string | null;
	salt?: string | null;
	hash?: string | null;
	_verified?: boolean | null;
	_verificationToken?: string | null;
	loginAttempts?: number | null;
	lockUntil?: string | null;
	password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
	id: string;
	prefix?: string | null;
	updatedAt: string;
	createdAt: string;
	url?: string | null;
	filename?: string | null;
	mimeType?: string | null;
	filesize?: number | null;
	width?: number | null;
	height?: number | null;
	focalX?: number | null;
	focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "guilds".
 */
export interface Guild {
	id: string;
	name: string;
	staff?: (string | User)[] | null;
	updatedAt: string;
	createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "submission-media".
 */
export interface SubmissionMedia {
	id: string;
	prefix?: string | null;
	updatedAt: string;
	createdAt: string;
	url?: string | null;
	filename?: string | null;
	mimeType?: string | null;
	filesize?: number | null;
	width?: number | null;
	height?: number | null;
	focalX?: number | null;
	focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "projects".
 */
export interface Project {
	id: string;
	title: string;
	slug: string;
	shortDescription: string;
	description: {
		root: {
			type: string;
			children: {
				type: string;
				version: number;
				[k: string]: unknown;
			}[];
			direction: ('ltr' | 'rtl') | null;
			format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
			indent: number;
			version: number;
		};
		[k: string]: unknown;
	};
	organizers: (string | Guild)[];
	status: 'draft' | 'hidden' | 'submissions-open' | 'submissions-closed' | 'ongoing' | 'past';
	links?:
		| {
				name: string;
				url: string;
				id?: string | null;
		  }[]
		| null;
	media?:
		| {
				type: 'image' | 'video';
				media?: string | Media | null;
				url?: string | null;
				id?: string | null;
		  }[]
		| null;
	date: string;
	image: string | Media;
	ogImage?: string | Media | null;
	'submission-url'?: string | null;
	collaborators?: (string | User)[] | null;
	hasSubmissions?: boolean | null;
	skin:
		| 'holoEN'
		| 'ina'
		| 'amelia'
		| 'gura'
		| 'kiara'
		| 'mori'
		| 'irys'
		| 'sana'
		| 'fauna'
		| 'kronii'
		| 'mumei'
		| 'baelz'
		| 'shiori'
		| 'bijou'
		| 'nerissa'
		| 'fuwawa'
		| 'mococo'
		| 'fuwamoco';
	flags?: (string | Flag)[] | null;
	devprops?:
		| {
				key: string;
				value: string;
				id?: string | null;
		  }[]
		| null;
	updatedAt: string;
	createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "flags".
 */
export interface Flag {
	id: string;
	code: string;
	name: string;
	updatedAt: string;
	createdAt: string;
}
// ...
export interface Submission {
	id: string;
	project: string | Project;
	author: string;
	srcIcon?: string | SubmissionMedia | null;
	message?: string | null;
	media?:
		| {
				type: 'image' | 'video';
				subtype?: ('artwork' | 'picture' | 'other') | null;
				image?: string | SubmissionMedia | null;
				url?: string | null;
				id?: string | null;
		  }[]
		| null;
	status: 'unchecked' | 'rejected' | 'accepted';
	filterableAttributes?:
		| {
				name: string;
				values?:
					| {
							value: string;
							id?: string | null;
					  }[]
					| null;
				id?: string | null;
		  }[]
		| null;
	devprops?:
		| {
				key: string;
				value: string;
				id?: string | null;
		  }[]
		| null;
	updatedAt: string;
	createdAt: string;
}